+++ frontend/src/views/ProfileView.vue
@@ -1,0 +1,266 @@
+<template>
+  <div class="layout__title">
+    <h1 class="title title--big">Мои данные</h1>
+  </div>
+
+  <div class="user">
+    <picture>
+      <img
+        :src="userStore.getWhoAmI.avatar"
+        :alt="userStore.getWhoAmI.name"
+        width="72"
+        height="72"
+      />
+    </picture>
+    <div class="user__name">
+      <span>{{ userStore.getWhoAmI.name }}</span>
+    </div>
+    <p class="user__phone">
+      Контактный телефон: <span>{{ userStore.getWhoAmI.phone }}</span>
+    </p>
+  </div>
+
+  <div
+    v-for="address in userStore.getAddresses"
+    :key="address.id"
+    class="layout__address"
+  >
+    <div class="sheet address-form">
+      <div class="address-form__header">
+        <b>{{ address.name }}</b>
+        <div class="address-form__edit">
+          <button type="button" class="icon">
+            <span class="visually-hidden">Изменить адрес</span>
+          </button>
+        </div>
+      </div>
+      <p>{{ address.fullAddress }}</p>
+      <small>{{ address.comment }}</small>
+    </div>
+  </div>
+
+  <div class="layout__address">
+    <form
+      action="test.html"
+      method="post"
+      class="address-form address-form--opened sheet"
+    >
+      <div class="address-form__header">
+        <b>Адрес №1</b>
+      </div>
+
+      <div class="address-form__wrapper">
+        <div class="address-form__input">
+          <label class="input">
+            <span>Название адреса*</span>
+            <input
+              type="text"
+              name="addr-name"
+              placeholder="Введите название адреса"
+              required
+            />
+          </label>
+        </div>
+        <div class="address-form__input address-form__input--size--normal">
+          <label class="input">
+            <span>Улица*</span>
+            <input
+              type="text"
+              name="addr-street"
+              placeholder="Введите название улицы"
+              required
+            />
+          </label>
+        </div>
+        <div class="address-form__input address-form__input--size--small">
+          <label class="input">
+            <span>Дом*</span>
+            <input
+              type="text"
+              name="addr-house"
+              placeholder="Введите номер дома"
+              required
+            />
+          </label>
+        </div>
+        <div class="address-form__input address-form__input--size--small">
+          <label class="input">
+            <span>Квартира</span>
+            <input
+              type="text"
+              name="addr-apartment"
+              placeholder="Введите № квартиры"
+            />
+          </label>
+        </div>
+        <div class="address-form__input">
+          <label class="input">
+            <span>Комментарий</span>
+            <input
+              type="text"
+              name="addr-comment"
+              placeholder="Введите комментарий"
+            />
+          </label>
+        </div>
+      </div>
+
+      <div class="address-form__buttons">
+        <button type="button" class="button button--transparent">
+          Удалить
+        </button>
+        <button type="submit" class="button">Сохранить</button>
+      </div>
+    </form>
+  </div>
+
+  <div class="layout__button">
+    <button type="button" class="button button--border">
+      Добавить новый адрес
+    </button>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { useUserStore } from "../stores/user";
+const userStore = useUserStore();
+</script>
+
+<style lang="scss" scoped>
+@import "@/assets/scss/app.scss";
+
+// user
+.user {
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+
+  margin-bottom: 33px;
+}
+
+.user__name {
+  @include b-s20-h23;
+
+  margin-left: 30px;
+
+  span {
+    display: inline-block;
+
+    vertical-align: middle;
+  }
+}
+
+.user__button {
+  display: inline-block;
+
+  cursor: pointer;
+  vertical-align: middle;
+}
+
+.user__phone {
+  @include b-s16-h19;
+
+  width: 100%;
+  margin-top: 20px;
+
+  span {
+    font-weight: 400;
+  }
+}
+
+.layout__address {
+  margin-top: 16px;
+}
+
+.address-form {
+  $bl: &;
+
+  position: relative;
+
+  padding-top: 0;
+  padding-bottom: 26px;
+
+  &--opened {
+    #{$bl}__header {
+      padding: 16px;
+    }
+  }
+
+  p {
+    @include r-s16-h19;
+
+    margin-top: 0;
+    margin-bottom: 16px;
+    padding: 0 16px;
+  }
+
+  small {
+    @include l-s11-h13;
+
+    display: block;
+
+    padding: 0 16px;
+  }
+}
+
+.address-form__wrapper {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+
+  width: 80%;
+  padding: 16px;
+}
+
+.address-form__input {
+  width: 100%;
+  margin-bottom: 16px;
+
+  &--size {
+    &--normal {
+      width: 60.5%;
+    }
+
+    &--small {
+      width: 18%;
+    }
+  }
+}
+
+.address-form__buttons {
+  display: flex;
+  justify-content: flex-end;
+
+  padding: 0 16px;
+
+  button {
+    margin-left: 16px;
+    padding: 16px 27px;
+  }
+}
+
+.address-form__header {
+  @include b-s14-h16;
+
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  margin-bottom: 21px;
+  padding: 10px 16px;
+
+  border-bottom: 1px solid rgba($green-500, 0.1);
+}
+
+.layout__button {
+  margin-top: 40px;
+
+  button {
+    padding: 12px 23px;
+  }
+}
+
+.layout__title {
+  margin-bottom: 27px;
+}
+</style>
